runner: Support test/test-* including non-js files
When falling back on the "is it executable" test for files that don't
end in .js or .coffee, files that are determined to be executable as
tests should be referenced by a more complete path because $CWD is
highly unlikely to be included in $PATH.

Update tests to expect full paths of commands.

Update package.json to run the existing test for this scenario.