Clean up stack traces
This removes a huge source of excessive noise.  You almost
never care about the fact that a stack trace originates in
Module._compile or where in timers.js the timeout was called
from.

Also removes the '    at ' from each line of the stack trace,
and the error name+message that is duplicated in the test point
anyway.